[FIX] project: sort kanban header/renderer functions
the deleteGroup method in the header opens the confirmation dialog then calls the kanbanrenderer deleteGroup method so no need to override it
the editGroup canEditGroup and canDeleteGroup methods must be placed and overrided in the kanbanHeader
as well as the onWillStart the sets the this.isProjectManager, which will be used in the canEditGroup
the canCreateGroup and isProjectTasksContext should stay in the renderer

Task-3276463
