+++ addons/project/static/src/views/project_task_kanban/project_task_kanban_header.js
@@ -1,9 +1,25 @@
 /** @odoo-module */
-
+import { useService } from '@web/core/utils/hooks';
 import { KanbanHeader } from "@web/views/kanban/kanban_header";
 import { FormViewDialog } from "@web/views/view_dialogs/form_view_dialog";
+import { onWillStart } from "@odoo/owl";
 
 export class ProjectTaskKanbanHeader extends KanbanHeader {
+    setup() {
+        super.setup();
+        this.action = useService('action');
+        this.userService = useService('user');
+
+        this.isProjectManager = false;
+        onWillStart(this.onWillStart);
+    }
+
+    async onWillStart() {
+        if (!this.props.list.isGroupedByPersonalStages) { // no need to check it if the group by is personal stages
+            this.isProjectManager = await this.userService.hasGroup('project.group_project_manager');
+        }
+    }
+
     editGroup() {
         const { resModel, value } = this.group;
         const groupBy = this.props.list.groupBy;
@@ -24,5 +40,27 @@ export class ProjectTaskKanbanHeader extends KanbanHeader {
                 this.props.list.model.notify();
             },
         });
+    }
+
+    async deleteGroup() {
+        if (this.group.groupByField.name === 'stage_id') {
+            const action = await this.group.model.orm.call(
+                this.group.resModel,
+                'unlink_wizard',
+                [this.group.resId],
+                { context: this.group.context },
+            );
+            this.action.doAction(action);
+            return;
+        }
+        super.deleteGroup();
+    }
+
+    canEditGroup(group) {
+        return super.canEditGroup(group) && (!this.props.list.isGroupedByStage || this.isProjectManager) || this.props.list.isGroupedByPersonalStages;
+    }
+
+    canDeleteGroup(group) {
+        return super.canDeleteGroup(group) && (!this.props.list.isGroupedByStage || this.isProjectManager) || this.props.list.isGroupedByPersonalStages;
     }
 }
