+++ addons/project/static/src/views/project_task_kanban/project_task_kanban_renderer.js
@@ -1,20 +1,15 @@
 /** @odoo-module */
 
-import { useService } from '@web/core/utils/hooks';
 import { KanbanRenderer } from '@web/views/kanban/kanban_renderer';
 import { ProjectTaskKanbanRecord } from './project_task_kanban_record';
 import { ProjectTaskKanbanHeader } from './project_task_kanban_header';
-
-const { onWillStart } = owl;
+import { useService } from '@web/core/utils/hooks';
 
 export class ProjectTaskKanbanRenderer extends KanbanRenderer {
     setup() {
         super.setup();
-        this.userService = useService('user');
         this.action = useService('action');
 
-        this.isProjectManager = false;
-        onWillStart(this.onWillStart);
     }
 
     get canMoveRecords() {
@@ -37,40 +32,12 @@ export class ProjectTaskKanbanRenderer extends KanbanRenderer {
         return canResequenceGroups;
     }
 
-    async onWillStart() {
-        if (!this.props.list.isGroupedByPersonalStages) { // no need to check it if the group by is personal stages
-            this.isProjectManager = await this.userService.hasGroup('project.group_project_manager');
-        }
-    }
-
-    isProjectTasksContext() {
-        return this.props.list.context.active_model === "project.project" && this.props.list.context.default_project_id;
-    }
-
     canCreateGroup() {
         return (super.canCreateGroup() && this.isProjectTasksContext() && this.props.list.isGroupedByStage) || this.props.list.isGroupedByPersonalStages;
     }
 
-    canDeleteGroup(group) {
-        return super.canDeleteGroup(group) && (!this.props.list.isGroupedByStage || this.isProjectManager) || this.props.list.isGroupedByPersonalStages;
-    }
-
-    canEditGroup(group) {
-        return super.canEditGroup(group) && (!this.props.list.isGroupedByStage || this.isProjectManager) || this.props.list.isGroupedByPersonalStages;
-    }
-
-    async deleteGroup(group) {
-        if (group && group.groupByField.name === 'stage_id') {
-            const action = await group.model.orm.call(
-                group.resModel,
-                'unlink_wizard',
-                [group.resId],
-                { context: group.context },
-            );
-            this.action.doAction(action);
-            return;
-        }
-        super.deleteGroup(group);
+    isProjectTasksContext() {
+        return this.props.list.context.active_model === "project.project" && this.props.list.context.default_project_id;
     }
 }
